requirements.txt
colorama==0.4.6
cupy-cuda12x==13.2.0
fastrlock==0.8.2
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
platformdirs==4.2.2
pycuda==2024.1.2
pytools==2024.1.12
tqdm==4.66.5
typing_extensions==4.12.2
